pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "youtuber2skill"
version = "0.1.0"
description = "Convert YouTube videos into structured AI Skills"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "yt-dlp>=2024.0",
    "openai>=1.0",
    "pywhispercpp>=1.0",
    "pyyaml>=6.0",
]

[project.scripts]
youtuber2skill = "youtuber2skill.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
